# RideHailMatching: Per thread GeodeticCalculator